Add local mode to Package
