+++ .gitignore
@@ -16,3 +16,6 @@ node_modules/
 
 # Dist
 dist
+
+./*/*.json
+./*/*.yml
