+++ localFiles.js
@@ -1,0 +1,5 @@
+const utils = require("./utils");
+
+["icons.yml", "categories.yml"].forEach((fileName) => {
+  utils.generateJSONFile("local_icons", fileName);
+});
