+++ webpack.config.js
@@ -1,6 +1,7 @@
 const path = require("path");
 const SRC_DIR = path.resolve(__dirname, "src");
 const SEARCH_ICONS = path.resolve(__dirname, "search_icons");
+const UTILS = path.resolve(__dirname, "utils");
 
 module.exports = {
   entry: "./index.js",
@@ -11,7 +12,8 @@ module.exports = {
   resolve: {
     alias: {
       Src: SRC_DIR,
-      SearchIcons: SEARCH_ICONS
+      SearchIcons: SEARCH_ICONS,
+      Utils: UTILS
     },
     extensions: [".js", ".json"]
   }
